requirements.txt
asgiref==3.2.3
beautifulsoup4==4.6.0
boto3==1.10.25
botocore==1.13.25
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coderedcms==0.16.3
decorator==4.4.1
dj-database-url==0.5.0
dj-static==0.0.6
django-bootstrap4==1.0.1
django-cogwheels==0.3
django-colorful==1.3
django-el-pagination==3.2.4
django-eventtools==0.9.11
django-heroku==0.3.1
django-modelcluster==5.0
django-social-share==1.3.2
django-storages==1.8
django-taggit==1.1.0
django-toolbelt==0.0.1
django-treebeard==4.3
django_sendmail_backend
djangorestframework==3.10.3
docutils==0.15.2
draftjs-exporter==2.1.7
future==0.18.2
geocoder==1.38.1
gunicorn==20.0.0
html5lib==1.0.1
icalendar==4.0.4
idna==2.8
invoke==1.3.0
jmespath==0.9.4
Pillow==6.2.1
psycopg2==2.8.4
puput==1.0.5
python-dateutil==2.8.1
pytz==2019.3
ratelim==0.1.6
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
tapioca-disqus==0.1.2
tapioca-wrapper==0.5.3
Unidecode==1.1.1
urllib3==1.25.7
wagtail-cache==0.5.2
wagtail-import-export==0.2
wagtail==2.7
wagtailfontawesome==1.1.4
wagtailmenus==3.0
webencodings==0.5.1
whitenoise==4.1.4
Willow==1.3
